// src/services/-auth-login-post.schemas.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Shopping Platform API
 * Production-grade e-commerce backend
 * OpenAPI spec version: 1.0
 */
export interface DtoLoginRequest {
  email: string;
  password: string;
}

export interface DtoUserResponse {
  email?: string;
  first_name?: string;
  id?: number;
  last_name?: string;
  phone?: string;
  role?: string;
}

export interface DtoLoginResponseData {
  access_token?: string;
  refresh_token?: string;
  user?: DtoUserResponse;
}

export interface DtoLoginResponse {
  data?: DtoLoginResponseData;
  message?: string;
  success?: boolean;
}

export interface DtoMessageResponse {
  message?: string;
  success?: boolean;
}